Split scheduling out into it's own sub components and elements
